// pharmacy-ui/src/app/services/models/cart-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CartProductResponse } from '../models/cart-product-response';
export interface CartResponse {
  id?: number;
  ordered?: boolean;
  product?: Array<CartProductResponse>;
  status?: 'STARTED' | 'CHECKED_OUT' | 'ABANDONED' | 'REMOVED';
  userId?: number;
}